backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "yifan",email = "jeromeyao1999@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "chromadb (>=1.5.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
